requirements.txt
arrow==1.3.0
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
Django==4.2.11
django-ckeditor==6.7.1
django-ckeditor-5==0.2.12
django-cors-headers==4.4.0
django-js-asset==2.2.0
django-light==0.1.0.post3
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
filelock==3.13.4
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mysqlclient==2.2.4
openapi-codec==1.3.2
packaging==24.0
pillow==10.3.0
platformdirs==4.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
soupsieve==2.6
sqlparse==0.4.4
types-python-dateutil==2.8.19.20240311
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.1
